+++ examples/resources/meraki_network_device_claim/resource.tf
@@ -1,4 +1,15 @@
 resource "meraki_network_device_claim" "example" {
   network_id = "123456"
-  serials    = ["1234-1234-1234"]
+  details_by_device = [
+    {
+      serial = "Q234-ABCD-5678"
+      details = [
+        {
+          name  = "username"
+          value = "milesmeraki"
+        }
+      ]
+    }
+  ]
+  serials = ["1234-1234-1234"]
 }