+++ internal/provider/data_source_meraki_switch_ports.go
@@ -103,6 +103,10 @@ func (d *SwitchPortsDataSource) Schema(ctx context.Context, req datasource.Schem
 							MarkdownDescription: "The link speed for the switch port.",
 							Computed:            true,
 						},
+						"mac_whitelist_limit": schema.Int64Attribute{
+							MarkdownDescription: "The maximum number of MAC addresses for regular MAC allow list. Only applicable when `accessPolicyType` is `MAC allow list`. Note: Config only supported on verions greater than ms18 only for classic switches.",
+							Computed:            true,
+						},
 						"name": schema.StringAttribute{
 							MarkdownDescription: "The name of the switch port.",
 							Computed:            true,